Check if the folder system/cache/dca exists before scanning a non existing folder
This prevents an error if the folder is no there (e.g. the cache is disabled).
Fix was taken from David Maack's pull request: davidmaack/check-if-the-folder-systemcachedca-exist-1382627538858

+++ system/modules/efg/classes/FormdataBackend.php
@@ -142,14 +142,17 @@ class FormdataBackend extends \Backend
 		}
 
 		// Remove cached dca files
-		$arrFiles = scan(TL_ROOT . '/system/cache/dca', true);
-
-		foreach ($arrFiles as $strFile)
+		if (is_dir(TL_ROOT . '/system/cache/dca'))
 		{
-			if (substr($strFile, 0, 3) == 'fd_' || $strFile == 'tl_formdata.php')
+			$arrFiles = scan(TL_ROOT . '/system/cache/dca', true);
+
+			foreach ($arrFiles as $strFile)
 			{
-				$objFile = new \File('system/cache/dca/' . $strFile);
-				$objFile->delete();
+				if (substr($strFile, 0, 3) == 'fd_' || $strFile == 'tl_formdata.php')
+				{
+					$objFile = new \File('system/cache/dca/' . $strFile);
+					$objFile->delete();
+				}
 			}
 		}
